+++ src/simulator/coordinate.rs
@@ -1,5 +1,5 @@
-pub fn get_coordinate(length: f64, nitems: usize, index: usize) -> f64 {
-    length * ((index as f64 + 0.5f64) / nitems as f64)
+pub fn get_coordinate(length: f32, nitems: usize, index: usize) -> f32 {
+    length * ((index as f32 + 0.5f32) / nitems as f32)
 }
 
 #[cfg(test)]
@@ -7,7 +7,7 @@ mod test {
     use super::get_coordinate;
     #[test]
     fn check() {
-        let length = 1f64;
+        let length = 1f32;
         let nitems = 4usize;
         assert!(0.125 == get_coordinate(length, nitems, 0));
         assert!(0.375 == get_coordinate(length, nitems, 1));
