+++ src/simulator/source.rs
@@ -3,21 +3,21 @@ use crate::simulator::Config;
 use crate::simulator::NDIMS;
 
 struct Quadratic {
-    a: f64,
-    b: f64,
-    c: f64,
+    a: f32,
+    b: f32,
+    c: f32,
 }
 
 pub struct Distribution {
-    pub mean: f64,
-    pub deviation: f64,
+    pub mean: f32,
+    pub deviation: f32,
 }
 
 pub struct Point {
-    pub position: f64,
-    interval: f64,
+    pub position: f32,
+    interval: f32,
     distribution: Distribution,
-    velocity: f64,
+    velocity: f32,
     quadratic: Quadratic,
 }
 
@@ -25,14 +25,14 @@ impl Point {
     pub fn new(
         rng: &mut Random,
         distribution: Distribution,
-        interval: f64,
-        position: f64,
-        velocity: f64,
+        interval: f32,
+        position: f32,
+        velocity: f32,
     ) -> Self {
         let mut quadratic = Quadratic {
-            a: 0f64,
-            b: 0f64,
-            c: 0f64,
+            a: 0f32,
+            b: 0f32,
+            c: 0f32,
         };
         update_quadratic(
             rng,
@@ -51,7 +51,7 @@ impl Point {
         }
     }
 
-    pub fn update(&mut self, rng: &mut Random, dt: f64) {
+    pub fn update(&mut self, rng: &mut Random, dt: f32) {
         update_quadratic(
             rng,
             &self.distribution,
@@ -60,25 +60,25 @@ impl Point {
             get_velocity(&self.quadratic, dt),
             &mut self.quadratic,
         );
-        self.position = get_position(&self.quadratic, 0f64);
-        self.velocity = get_velocity(&self.quadratic, 0f64);
+        self.position = get_position(&self.quadratic, 0f32);
+        self.velocity = get_velocity(&self.quadratic, 0f32);
     }
 }
 
-fn get_position(q: &Quadratic, time: f64) -> f64 {
+fn get_position(q: &Quadratic, time: f32) -> f32 {
     q.a * time.powi(2) + q.b * time + q.c
 }
 
-fn get_velocity(q: &Quadratic, time: f64) -> f64 {
-    2f64 * q.a * time + q.b
+fn get_velocity(q: &Quadratic, time: f32) -> f32 {
+    2f32 * q.a * time + q.b
 }
 
 fn update_quadratic(
     rng: &mut Random,
     distribution: &Distribution,
-    t_aft: f64,
-    x_bef: f64,
-    u_bef: f64,
+    t_aft: f32,
+    x_bef: f32,
+    u_bef: f32,
     q: &mut Quadratic,
 ) {
     // determine
@@ -96,9 +96,9 @@ fn update_quadratic(
     // | t_aft^2   t_aft   1 | | b | = | x_aft |
     // |       0       1   0 | | c | = | u_bef |
     let x_aft = {
-        let mean: f64 = distribution.mean;
-        let dev: f64 = distribution.deviation;
-        mean + rng.gen_range(0f64 - dev, 0f64 + dev)
+        let mean: f32 = distribution.mean;
+        let dev: f32 = distribution.deviation;
+        mean + rng.gen_range(0f32 - dev, 0f32 + dev)
     };
     q.a = (x_aft - x_bef - u_bef * t_aft) / t_aft.powi(2);
     q.b = u_bef;
@@ -107,23 +107,23 @@ fn update_quadratic(
 
 pub struct Source {
     pub points: [Point; NDIMS],
-    pub amp: f64,
-    omega: f64,
-    time: f64,
+    pub amp: f32,
+    omega: f32,
+    time: f32,
 }
 
 impl Source {
     pub fn new(config: &Config, rng: &mut Random) -> Self {
-        let position = [0.5f64 * config.lengths[0], 0.5f64 * config.lengths[1]];
-        let velocity = [0f64, 0f64];
+        let position = [0.5f32 * config.lengths[0], 0.5f32 * config.lengths[1]];
+        let velocity = [0f32, 0f32];
         let points = [
             Point::new(
                 rng,
                 Distribution {
                     mean: position[0],
-                    deviation: 16f64 * config.lengths[0],
+                    deviation: 16f32 * config.lengths[0],
                 },
-                25f64,
+                25f32,
                 position[0],
                 velocity[0],
             ),
@@ -131,16 +131,16 @@ impl Source {
                 rng,
                 Distribution {
                     mean: position[1],
-                    deviation: 16f64 * config.lengths[1],
+                    deviation: 16f32 * config.lengths[1],
                 },
-                25f64,
+                25f32,
                 position[1],
                 velocity[1],
             ),
         ];
-        let amp = 1f64;
-        let omega = 1.5f64 * std::f64::consts::PI;
-        let time = 0f64;
+        let amp = 1f32;
+        let omega = 1.5f32 * std::f32::consts::PI;
+        let time = 0f32;
         Self {
             points,
             amp,
@@ -149,11 +149,11 @@ impl Source {
         }
     }
 
-    pub fn get_position(&self) -> [f64; NDIMS] {
+    pub fn get_position(&self) -> [f32; NDIMS] {
         [self.points[0].position, self.points[1].position]
     }
 
-    pub fn update(&mut self, rng: &mut Random, dt: f64) {
+    pub fn update(&mut self, rng: &mut Random, dt: f32) {
         self.points[0].update(rng, dt);
         self.points[1].update(rng, dt);
         self.time += dt;
